+++ tests/testthat/test_options.R
@@ -3,8 +3,8 @@ library(shinystan)
 context("launch options")
 
 test_that("options set when package loads", {
-  expect_that(getOption("shinystan.rstudio"), is_false())
+  expect_false(getOption("shinystan.rstudio"))
   options(shinystan.rstudio = TRUE)
-  expect_that(getOption("shinystan.rstudio"), is_true())
+  expect_true(getOption("shinystan.rstudio"))
   options(shinystan.rstudio = FALSE)
 })