+++ NEWS.md
@@ -8,6 +8,7 @@ parameters).
 can be called.
 * Allow selecting `yrep` from global environment for PPcheck
 * Change `as.shinystan` to S4 generic with methods
+* Add **rsconnect** to `Imports` in `DESCRIPTION`.
 
 ### Version 2.1.0
 * Compatibility with recent ggplot2 update